Noetic_seed/requirements.txt
httpx
psutil
numpy
sqlalchemy
aiosqlite
onnxruntime
tokenizers
huggingface-hub
playwright
Pillow
lmstudio>=1.1.0
